requirements.txt
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
flask-core==2.9.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
